Updated TODOs and repacking HE0 should now work properly

+++ HumongousResourceEditor/src/low_level/HumongousHeader.cpp
@@ -8,13 +8,13 @@ namespace resource_editor
 {
 	namespace chunk_reader
 	{
-        HumongousHeader::HumongousHeader(const char* a_ChunkID, size_t a_ChunkSize, bool a_BigEndian)
+        HumongousHeader::HumongousHeader(const char* a_ChunkID, uint32_t a_ChunkSize, bool a_BigEndian)
         {
 			memcpy(chunk_id, a_ChunkID, CHUNK_ID_SIZE);
 			SetChunkSize(a_ChunkSize, a_BigEndian);
         }
 
-        HumongousHeader::HumongousHeader(unsigned char* a_ChunkID, size_t a_ChunkSize, bool a_BigEndian)
+        HumongousHeader::HumongousHeader(unsigned char* a_ChunkID, uint32_t a_ChunkSize, bool a_BigEndian)
         {
 			memcpy(chunk_id, a_ChunkID, CHUNK_ID_SIZE);
 			SetChunkSize(a_ChunkSize, a_BigEndian);
